Add search and figure stage viewing
